requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
fastapi==0.75.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.34
sqlalchemy2-stubs==0.0.2a21
sqlmodel==0.0.6
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
Werkzeug==2.0.3
